requirements.txt
sqlalchemy>=1.3.10
urllib3>=1.25.6
Pandas>=0.25.1
pyodbc>=4.0.27
python-dotenv>=0.10.3
